create static method into appointment model

+++ test/appointment.js
@@ -13,23 +13,32 @@ describe("************ Test Model Appointment", ()=>{
 				user: "592d4cc48967d7a92f030b87",
 				professional: "59319cc254ae912105ae03d2",
 				service: "59326bb3bc42cebc293ba164",
-				date: new Date()
+				date: new Date("2017/06/23 18:00:00")
 			}
 			request.post("/appointment")
-						.expect(201)
 						.set("authorization", "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI1OTMxNjVlOWJjNDJjZWJjMjkzYmExNjIiLCJsZXZlbCI6NSwidHlwZSI6ImFkbWluIiwiaWF0IjoxNDk2NDIyMDAzfQ.q29ExhIohwUoBpNzyEBHxqeK9i0GuSa27Aoq6kx0k6c")
 						.send(appointment)
 						.then((res) => {
-							expect(res.status).to.equal(201);
-							expect(res.body).to.have.property("_id");
-							expect(res.body).to.have.property("completed");
-							expect(res.body.completed).to.have.property("status", false);
-							console.log("Appointment created successfully .......")
-							return request.delete("/appointment/" + res.body._id)
-														.expect(200)
-														.set("authorization", "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI1OTMxNjVlOWJjNDJjZWJjMjkzYmExNjIiLCJsZXZlbCI6NSwidHlwZSI6ImFkbWluIiwiaWF0IjoxNDk2NDIyMDAzfQ.q29ExhIohwUoBpNzyEBHxqeK9i0GuSa27Aoq6kx0k6c")
+							let status = res.status
+							if(status === 201) {
+								expect(res.status).to.equal(201);
+								expect(res.body).to.have.property("_id");
+								expect(res.body).to.have.property("completed");
+								expect(res.body.completed).to.have.property("status", false);
+								console.log("Appointment created successfully .......")
+								return request.delete("/appointment/" + res.body._id)
+															.expect(200)
+															.set("authorization", "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiI1OTMxNjVlOWJjNDJjZWJjMjkzYmExNjIiLCJsZXZlbCI6NSwidHlwZSI6ImFkbWluIiwiaWF0IjoxNDk2NDIyMDAzfQ.q29ExhIohwUoBpNzyEBHxqeK9i0GuSa27Aoq6kx0k6c")
+							}else if(status === 409){
+								expect(res.status).to.equal(409);
+								expect(res.body).to.have.property("error");
+								expect(res.body).to.have.property("message");
+								done()
+							}
+
 						}, error => done(error))
 						.then((res)=>{
+								if(!res) return 
 								expect(res.status).to.equal(200)
 								expect(res.body).to.be.empty;
 								console.log("Appointment deleted successfully .......")
